multi-bay support: adding different bays
implements support for multiple different groups of specifications (for instance, if you have one bay as a headshot and another as a halfbody, you can count this now!)

button for ADDING has been implemented, and the dropdown lists will update properly. however the switching and removing functionality are still to come.

additional things:
- fixed the top row of buttons being misaligned
- removed some commented-out code to clean up a bit
- fixed bug with the counting string for crafts (the math was still working just fine tho)