hello info-manager
- created info-manager.js
- added info section to v2 (actually did that in the last commit oops) but it doesn't do anything yet
- info-manager holds the info object for now, which has a helpful little info element reference but more importantly...
- started describing the fuck out of everything in the info object. the idea is to eventually use the info block in the site to give the user information about anything they might need clarification on!
that will either be done with an "inspect" search glass if i can figure that out, or just. info buttons or something.
it will probably be implemented in stackable values first as those have like, things that are technically but not actually buttons that make me wanna click 'em but they don't do anything